Move code, this makes the upcoming commit easier to review
